move existing catalog/utils test to a location where they will be run
